updated CRUD series view count files

+++ netflix/src/main/java/com/example/netflix/repository/crud/MovieViewCountCrudRepository.java
@@ -4,7 +4,7 @@ import com.example.netflix.entity.MovieViewCount;
 import org.springframework.data.jpa.repository.query.Procedure;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MovieViewCountCrudRepository extends CrudRepository<MovieViewCount, Long> {
+public interface MovieViewCountCrudRepository extends CrudRepository<MovieViewCount, Integer> {
     @Procedure(name = "AddMovieViewCount")
     void addMovieViewCount(Integer movieId, Integer accountId);
 }
